#78: the epsilon tie-breaks ran on self.dtype and INVERTED in every f32 lane
Found while hunting the envoy divergence, by inspecting the GPU's worked-tile
pick against city.ts. It is NOT that bug — see below — but it is a live one.

THE BUG. The GPU broke worked-tile ties by perturbing the score,
    score = score - tc.to(self.dtype) * 1e-9   # tie: lowest index first
where TS sorts `b.score - a.score || a.index - b.index`: exact score, then
LOWEST index. Measured rather than reasoned about:

    f64  keys 39.9999991 / 39.999999995  -> picks the LOW index   (matches TS)
    f32  keys 40.0       / 40.0          -> picks the HIGH index   (inverted)

A 1e-9 index epsilon sits far below the f32 ULP of a score around 40 (~4e-6),
so in f32 it rounds away COMPLETELY and topk resolves exact ties by its own
unspecified order. `_auto_pick` carried the same construct with a 1e-6 epsilon
on tech/civic costs of several thousand beakers — even further below the ULP —
so equal-cost techs resolved by argmin order instead of table order.

WHY NO GATE EVER SAW IT. Every gate runs f64, where the construct is correct.
f32 is used by eval.py, behavior_probe.py, gen_targets.py and duel_eval.py —
the environment P8 will TRAIN in. An agent trained there would have worked
different tiles than the spec, permanently and invisibly. This is the
wrong-constant failure mode one level down: a gate-green divergence that only
the RL target ever experiences.

THE FIX, both sites: force the tie-break key to f64 with .double(), exactly as
the rival twin _rival_city_yields_all already does — that path was hardened
long ago and only the player path was left riding self.dtype. .double() is a
no-op on an f64 tensor, so no f64 gate number can move, and nothing downstream
takes dtype from the key (top_idx is long, top_scores feeds only a bool mask).

REACHABILITY, MEASURED, because the first version of the regression assert was
decoration. At the 30 turns the f32 poke already ran, f32 and f64 agree EVEN
WITH THE BUG PRESENT. Re-reverting the fix and sweeping fixtures x turn counts
put the divergence at 120 turns: seed9002 diverges in pop, seed9014 in pop AND
techs. With the fix seed9014 goes fully clean. That flip is the proof the bug
was live, and the poke lane now asserts exactly it.

RESIDUAL, RECORDED AND OPEN: seed9002 STILL diverges in pop at 120 turns after
the fix, so a SECOND dtype-dependent source exists. It may well be benign — f32
rounding flipping a growth threshold is expected, not a bug — but it is
unproven, so the lane does not assert it and the AUDIT carries it as open.

AND IT DOES NOT EXPLAIN THE ENVOY GAP, which is an f64 divergence. Ruled out by
the same measurement that found the bug: in f64 the construct orders exactly as
TS unless two tiles' true scores differ by a nonzero amount under ~1.1e-6, and
Civ tile scores move in halves. The previous commit's claim that this was the
root cause is corrected in the AUDIT; that hunt stands open.

Gates: scripted parity 24 x 250 0.0 milli (f64 unchanged, as predicted);
district-breadth poke lane green including the new 120-turn f32/f64 assert.

+++ gpu/civ6gpu/engine.py
@@ -1951,9 +1951,13 @@ class BatchSim:
         """Cheapest-available (effective cost, tie = table order), where cur == -1."""
         avail = self._available_mask(done, prereq)
         eff = self._eff_cost(cost.unsqueeze(0).expand_as(avail), boosted)
-        key = torch.where(avail, eff, torch.tensor(float("inf"), dtype=self.dtype, device=self.device))
-        # stable tie-break on index: add a tiny index epsilon
-        key = key + torch.arange(key.shape[1], device=self.device, dtype=self.dtype) * 1e-6
+        key = torch.where(avail, eff, torch.tensor(float("inf"), dtype=self.dtype, device=self.device)).double()
+        # stable tie-break on index: add a tiny index epsilon. #78: FORCED f64
+        # for the same reason as the worked-tile pick — a 1e-6 epsilon is below
+        # the f32 ULP of a several-thousand-beaker cost, so on self.dtype=f32 it
+        # rounded away and equal-cost techs/civics resolved by argmin's own
+        # order instead of table order. f64 lanes are unchanged.
+        key = key + torch.arange(key.shape[1], device=self.device, dtype=torch.float64) * 1e-6
         best = key.argmin(dim=1)
         has = avail.any(dim=1)
         return torch.where((cur == -1) & has, best, cur)
@@ -3193,8 +3197,19 @@ class BatchSim:
             # tile here while TS worked a real one.
             & (self.built_wonder.gather(1, tcf).reshape(B, C, M) < 0)
         )  # [B, C, M]
-        score = torch.where(cand, tile_score.gather(1, tcf).reshape(B, C, M), torch.tensor(-1e18, dtype=self.dtype, device=dev))
-        score = score - tc.to(self.dtype) * 1e-9  # tie: lowest index first
+        # #78: the tie-break runs in FORCED f64, like the rival twin
+        # (_rival_city_yields_all builds its key with an explicit .double()).
+        # Riding self.dtype silently BROKE this in the f32 lanes: an index
+        # epsilon of 1e-9 is far below the f32 ULP of a score around 40
+        # (~4e-6), so it rounded away completely and topk resolved exact ties
+        # by its own unspecified order — picking the HIGHEST index where TS
+        # (city.ts, `b.score - a.score || a.index - b.index`) takes the
+        # lowest. f64 lanes are arithmetically unchanged by this (.double()
+        # is a no-op on an f64 tensor), so no gate number moves; the f32 RL
+        # path (eval.py / behavior_probe.py / gen_targets.py / duel_eval.py)
+        # stops working different tiles than the spec.
+        score = torch.where(cand, tile_score.gather(1, tcf).reshape(B, C, M).double(), torch.tensor(-1e18, dtype=torch.float64, device=dev))
+        score = score - tc.double() * 1e-9  # tie: lowest index first
         k = min(max(int(self.pop.max().item()), 1), M)
         top_scores, top_idx = score.topk(k, dim=2)
         take = (torch.arange(k, device=dev).view(1, 1, k) < self.pop.unsqueeze(2)) & (top_scores > -1e17)
